requirements.txt
aiohttp==1.0.5
aiohttp-jinja2==0.13.0
appdirs==1.4.3
async-timeout==1.2.1
cffi==1.10.0
chardet==3.0.4
click==6.7
discord.py>=0.16.11
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.9.6
markovify==0.6.0
MarkupSafe==1.0
meteor-ejson==1.1.0
multidict==3.1.3
numexpr==2.6.2
numpy==1.12.1
packaging==16.8
pdfminer.six==20170419
psutil==5.2.2
pycparser==2.18
pycrypto==2.6.1
pyee==3.0.3
PyNaCl==1.0.1
pyparsing==2.2.0
python-ddp==0.1.5
redis==2.10.5
six==1.10.0
Unidecode==0.4.21
websockets==3.3
Werkzeug==0.12.2
ws4py==0.4.2
yarl==0.10.0
youtube-dl==2017.7.15
